pyproject.toml
[project]
name = "learn-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "logfire>=4.16.0",
    "pydantic-ai>=1.39.1",
    "pydantic-ai-slim[openai,openrouter]>=1.39.1",
    "python-dotenv>=1.2.1",
]
